refactor cfg code to app/cfg package
